Move edit measure functions into utils so that it can be shared
